backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
clickhouse-connect==0.6.22
python-dotenv==1.0.0
pydantic[email]==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
mcp>=1.0.0
